feat!(draganddrop): Add Drag and Drop logic to Backend
Add a move route for dragging and dropping cards to different section.

Card entity and database are modified to accept a order index number within a section

Refactor App.tsx for better readability and reusability

Reordering cards within a section has been added to the frontend and backend, but a bug still persists that prevents correct ordering in the front end.